updated basic player to vite project
