+++ examples/basic-player/index.html
@@ -1,12 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Oddysee HLS Audio Player - Beautiful Demo</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Oddysee HLS Audio Player Demo</title>
+  </head>
+  <body>
     <div class="container">
         <h1>🎵 Oddysee HLS Audio Player</h1>
         <p>Experience our beautiful API that wraps verbose HLS.js complexity into elegant audio controls</p>
@@ -61,6 +60,7 @@
         </div>
     </div>
 
-    <script type="module" src="app.js"></script>
-</body>
+    <!-- Vite entry point -->
+    <script type="module" src="/main.js"></script>
+  </body>
 </html>
